Add files via upload
For the first stab at the Wildcard project, I was curious about the possibility of performing a surrogate optimization on the pipeline I built, and how that would compare with the Bayesian optimization. In addition, I looked at the hyperparameter search space using parallel coordinate plots in order to better understand how it might be exploited in the future.